requirements.txt
fastmcp
httpx
pydantic
python-dotenv
nest-asyncio
